+++ internal/lib/nhostclient/storage/storage.go
@@ -1,0 +1,2 @@
+//go:generate oapi-codegen -generate types,client -response-type-suffix R -package storage -o storage.gen.go ../../../../services/storage/controller/openapi.yaml
+package storage